refactored eventsouring api
