--- pkg/apps/ixservice/domain/enum/go.mod
@@ -1,1 +1,0 @@
-module github.com/innoxchain/ixstorage/pkg/apps/ixservice/domain/enum
