+++ pkg/apps/ixservice/event/command_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/satori/go.uuid"
-	"github.com/innoxchain/ixstorage/pkg/apps/ixservice/domain/enum"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -16,7 +15,7 @@ func TestOrderCreatedCommand(t *testing.T) {
 
 	createCommand := CreateOrderCommand{
 		UUID: uuid.NewV4().String(),
-		Capacity: enum.ThreeGB,
+		Capacity: ThreeGB,
 	}
 
 	ApplyCommand(createCommand, &order)
@@ -24,5 +23,5 @@ func TestOrderCreatedCommand(t *testing.T) {
 	t.Log("Order UUID: ", order.GetAggregateID())
 
 	assert.True(t, order.UUID!="", "AggregateID must not be nil")
-	assert.Equal(t, enum.ThreeGB, order.Capacity, "Capacity must be ThreeGB after command has been applied")
+	assert.Equal(t, ThreeGB, order.Capacity, "Capacity must be ThreeGB after command has been applied")
 }
