+++ pkg/apps/ixservice/examples/eventsourcing/dbconfig/local_db.json
@@ -1,0 +1,5 @@
+{
+    "ConnectString": "postgresql://test@localhost:26257",
+    "DatabaseName" : "eventstore",
+    "Sslmode" : "disable"
+}
